+++ findiff/current/__init__.py
@@ -1,0 +1,3 @@
+API_VERSION = 1
+
+from .operators import *