+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## Version 0.10.x
+
+- Create symbolic representations of finite difference schemes
+
+## Version 0.9.x
+
+- Generate differential operators for generic stencils
+
 ## Version 0.8.x
 
 - Formulate and solve partial differential equations